+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `polygon` tag.
 - Allow specifying block height in the `src` string of `Widget` to lock a child component to a specific version i.e. `src="<path>@<block height>"`
 - Support custom React tags. Here is an example of how to use it:
+
 ```jsx
 const BgDiv = ({ children, color }) => (
   <div style={{ background: color }}>{children}</div>
@@ -12,6 +13,8 @@ const BgDiv = ({ children, color }) => (
 
 return <BgDiv color="red">Hello World</BgDiv>;
 ```
+
+- Add [nanoid](https://github.com/ai/nanoid#usage) support (available methods: `nanoid.nanoid()` and `nanoid.customAlphabet()`)
 
 ## 1.0.0